# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Outpoint": {
				"type": "string"
			},
...
			"TxLog": {
				"properties": {
					"txid": {
						"type": "string"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"idx": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"txid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OpnsResponse": {
				"properties": {
					"outpoint": {
						"$ref": "#/components/schemas/Outpoint"
					},
					"origin": {
						"$ref": "#/components/schemas/Outpoint"
					},
					"owner": {
						"type": "string"
					},
					"domain": {
						"type": "string"
					},
					"map": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"outpoint",
					"origin",
					"owner",
					"domain"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OpnsMine": {
				"properties": {
					"outpoint": {
						"$ref": "#/components/schemas/Outpoint"
					},
					"origin": {
						"$ref": "#/components/schemas/Outpoint"
					},
					"domain": {
						"type": "string"
					},
					"pow": {
						"type": "string"
					},
					"script": {
						"type": "string"
					}
				},
				"required": [
					"outpoint",
					"origin",
					"domain",
					"pow",
					"script"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "1sat-server-js",
		"contact": {}
	},
	"paths": {
...
		"/api/tx/bin": {
			"post": {
				"operationId": "BroadcastBuf",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "format",
						"required": false,
						"schema": {
							"default": "tx",
							"type": "string",
							"enum": [
								"tx",
								"ef",
								"beef"
							]
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string",
								"format": "byte"
							}
						}
					}
				}
			}
		},
		"/api/tx": {
			"post": {
				"operationId": "Broadcast",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "format",
						"required": false,
						"schema": {
							"default": "tx",
							"type": "string",
							"enum": [
								"tx",
								"ef",
								"beef"
							]
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"rawtx": {
										"type": "string"
									}
								},
								"required": [
									"rawtx"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/tx/{txid}/submit": {
			"get": {
				"operationId": "GetTxSubmit",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostTxSubmit",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/tx/{txid}/status": {
			"get": {
				"operationId": "GetTxStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"hash": {
											"type": "string"
										},
										"idx": {
											"type": "number",
											"format": "double"
										},
										"height": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"hash",
										"idx",
										"height"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/tx/{txid}/rawtx": {
			"get": {
				"operationId": "GetTxRawtx",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/tx/{txid}/proof": {
			"get": {
				"operationId": "GetTxProof",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/tx/{txid}": {
			"get": {
				"operationId": "GetTx",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/tx/batch": {
			"get": {
				"operationId": "GetTxBatch",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "txids",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				]
			},
			"post": {
				"operationId": "PostTxBatch",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/api/tx/address/{address}/from/{blockHeight}": {
			"get": {
				"operationId": "GetAddressTxids",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TxLog"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "blockHeight",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/tx/address/{address}/{txid}": {
			"post": {
				"operationId": "SaveTxPost",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "format",
						"required": false,
						"schema": {
							"default": "tx",
							"type": "string",
							"enum": [
								"tx",
								"ef",
								"beef"
							]
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"default": [],
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "broadcast",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"type": "string",
								"format": "byte"
							}
						}
					}
				}
			}
		},
...
		"/api/opns/{domain}": {
			"get": {
				"operationId": "GetOpns",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpnsResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "domain",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/opns/{domain}/mine": {
			"get": {
				"operationId": "GetOpnsMine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OpnsMine"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "domain",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}